Refactor some testing and github actions jobs.
