+++ .github/workflows/config.yml
@@ -1,64 +1,63 @@
-# This is a basic workflow to help you get started with Actions
-
 name: ci
-
-# Controls when the action will run. 
 on: [push]
-
-# A workflow run is made up of one or more jobs that can run sequentially or in parallel
 jobs:
-  # This workflow contains a single job called "build"
   build:
-    # The type of runner that the job will run on
     runs-on: ubuntu-latest
     container:
       image: docker://nrel/openstudio:3.1.0
-
-    # Steps represent a sequence of tasks that will be executed as part of the job
     steps:
-      # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
       - uses: actions/checkout@v2
-
       - name: Install gems
         run: |
           rm -f Gemfile.lock && bundle install
-
-      - name: Run all integrity checks
+      - name: Run integrity checks for all projects
         run: |
-          rake test:integrity_checks
-
-      - name: Run all unit tests
+          rake unit_tests:project_integrity_checks
+      - name: Run all integrity check unit tests
         run: |
-          rake test:unit_tests
-
-      - name: Run measures osw test
+          rake unit_tests:integrity_check_tests
+      - name: Run all measure tests
         run: |
-          rake test:measures_osw
+          rake unit_tests:measure_tests
 
-      - name: Run all regression tests
+  integration:
+    runs-on: ubuntu-latest
+    needs: build
+    container:
+      image: docker://nrel/openstudio:3.1.0
+    steps:
+      - uses: actions/checkout@v2
+      - name: Run integration tests for sampled datapoints
         run: |
-          rake test:regression_tests
-
-      - name: Store example osws
-        uses: actions/upload-artifact@v2
-        with:
-          path: workflows/results
-          name: example_osws          
-
-      - name: Store measures osws
+          rake integration:workflow_tests
+      - name: Store workflow results
         uses: actions/upload-artifact@v2
         with:
           path: |
-            test/test_measures_osw/buildstock.csv
-            test/test_measures_osw/results/results.csv
-            test/test_measures_osw/osw
-            test/test_measures_osw/xml
-          name: measures_osw
+            test/test_samples_osw/baseline
+            test/test_samples_osw/upgrades
+          name: samples
+
+  regression:
+    runs-on: ubuntu-latest
+    needs: build
+    container:
+      image: docker://nrel/openstudio:3.1.0
+    steps:
+      - uses: actions/checkout@v2
+      - name: Run regression tests for all example osws
+        run: |
+          rake regression:workflow_tests
+      - name: Store workflow results
+        uses: actions/upload-artifact@v2
+        with:
+          path: workflows/results
+          name: examples 
 
   docs:
     # The type of runner that the job will run on
     runs-on: ubuntu-latest
-    needs: build
+    needs: [integration, regression]
     strategy:
       matrix:
         python-version: [3.8.4]
