+++ .gitignore
@@ -7,12 +7,11 @@
 /project_testing/housing_characteristics/buildstock.csv
 /project_testing/outputs
 /lib
-/test/test_measures_osw/buildstock.csv
-/test/test_measures_osw/generated_files
-/test/test_measures_osw/osw
-/test/test_measures_osw/xml
-/test/test_measures_osw/results
-/test/test_measures_osw/out.osw
+/test/test_samples_osw/buildstock.csv
+/test/test_samples_osw/generated_files
+/test/test_measures_osw/baseline
+/test/test_measures_osw/upgrades
+/test/test_samples_osw/out.osw
 /resources/buildstock.csv
 .vscode
 /weather/*.ddy
