+++ Rakefile
@@ -8,9 +8,9 @@ require 'rake/testtask'
 require 'ci/reporter/rake/minitest'
 
 desc 'Perform tasks related to unit tests'
-namespace :test do
+namespace :unit_tests do
   desc 'Run integrity checks for all projects'
-  Rake::TestTask.new('integrity_checks') do |t|
+  Rake::TestTask.new('project_integrity_checks') do |t|
     t.libs << 'test'
     t.test_files = Dir['project_*/tests/*.rb']
     t.warning = false
@@ -18,23 +18,37 @@ namespace :test do
   end
 
   desc 'Run all integrity check unit tests'
-  Rake::TestTask.new('unit_tests') do |t|
+  Rake::TestTask.new('integrity_check_tests') do |t|
     t.libs << 'test'
-    t.test_files = Dir['test/test_integrity_checks.rb'] + Dir['measures/*/tests/*.rb']
+    t.test_files = Dir['test/test_integrity_checks.rb']
     t.warning = false
     t.verbose = true
   end
 
-  desc 'Run measures osw test for a sampled datapoint'
-  Rake::TestTask.new('measures_osw') do |t|
+  desc 'Run all measure tests'
+  Rake::TestTask.new('measure_tests') do |t|
     t.libs << 'test'
-    t.test_files = Dir['test/test_measures_osw.rb']
+    t.test_files = Dir['measures/*/tests/*.rb']
     t.warning = false
     t.verbose = true
   end
+end
 
+desc 'Perform tasks related to integration tests'
+namespace :integration do
+  desc 'Run integration tests for sampled datapoints'
+  Rake::TestTask.new('workflow_tests') do |t|
+    t.libs << 'test'
+    t.test_files = Dir['test/test_samples.rb']
+    t.warning = false
+    t.verbose = true
+  end
+end
+
+desc 'Perform tasks related to regression tests'
+namespace :regression do
   desc 'Run regression tests for all example osws'
-  Rake::TestTask.new('regression_tests') do |t|
+  Rake::TestTask.new('workflow_tests') do |t|
     t.libs << 'test'
     t.test_files = Dir['workflows/tests/*.rb']
     t.warning = false
