+++ measures/UpgradeCosts/measure.xml
@@ -3,8 +3,8 @@
   <schema_version>3.0</schema_version>
   <name>upgrade_costs</name>
   <uid>ef51212c-acc4-48d7-9b29-cf2a5c6c4449</uid>
-  <version_id>b6371e02-2094-4399-aa3d-6a64f41dd567</version_id>
-  <version_modified>20210203T194712Z</version_modified>
+  <version_id>710b5e2c-f1fe-4a4c-9e13-d0d0391edfde</version_id>
+  <version_modified>20210208T233621Z</version_modified>
   <xml_checksum>9BF1E6AC</xml_checksum>
   <class_name>UpgradeCosts</class_name>
   <display_name>Upgrade Costs</display_name>
@@ -176,30 +176,6 @@
       <filetype>osw</filetype>
       <usage_type>test</usage_type>
       <checksum>88052D64</checksum>
-    </file>
-    <file>
-      <filename>in.epw</filename>
-      <filetype>epw</filetype>
-      <usage_type>test</usage_type>
-      <checksum>E23378AA</checksum>
-    </file>
-    <file>
-      <filename>in.osm</filename>
-      <filetype>osm</filetype>
-      <usage_type>test</usage_type>
-      <checksum>AB0F7788</checksum>
-    </file>
-    <file>
-      <filename>in.xml</filename>
-      <filetype>xml</filetype>
-      <usage_type>test</usage_type>
-      <checksum>51CAFFD0</checksum>
-    </file>
-    <file>
-      <filename>SFD_1story_FB_UA_GRG_MSHP_FuelTanklessWH.xml</filename>
-      <filetype>xml</filetype>
-      <usage_type>test</usage_type>
-      <checksum>FC4F3075</checksum>
     </file>
   </files>
 </measure>
