+++ test/test_samples_osw/run/eplusout.end
@@ -1,0 +1,1 @@
+EnergyPlus Completed Successfully-- 163186 Warning; 0 Severe Errors; Elapsed Time=00hr 00min  6.80sec
