+++ test/test_samples_osw/run/finished.job
@@ -1,0 +1,1 @@
+Finished Workflow 2021-02-08 16:28:49 -0700
