+++ test/test_samples_osw/run/objectives.json
@@ -1,0 +1,2 @@
+{
+}
