+++ test/test_samples_osw/run/results_annual.csv
@@ -1,0 +1,149 @@
+Fuel Use: Electricity: Total (MBtu),25.53
+Fuel Use: Electricity: Net (MBtu),25.53
+Fuel Use: Natural Gas: Total (MBtu),0.0
+Fuel Use: Fuel Oil: Total (MBtu),1.21
+Fuel Use: Propane: Total (MBtu),0.0
+Fuel Use: Wood Cord: Total (MBtu),0.0
+Fuel Use: Wood Pellets: Total (MBtu),0.0
+Fuel Use: Coal: Total (MBtu),0.0
+
+End Use: Electricity: Heating (MBtu),0.34
+End Use: Electricity: Heating Fans/Pumps (MBtu),0.01
+End Use: Electricity: Cooling (MBtu),11.46
+End Use: Electricity: Cooling Fans/Pumps (MBtu),1.1
+End Use: Electricity: Hot Water (MBtu),1.12
+End Use: Electricity: Hot Water Recirc Pump (MBtu),0.0
+End Use: Electricity: Hot Water Solar Thermal Pump (MBtu),0.0
+End Use: Electricity: Lighting Interior (MBtu),0.0
+End Use: Electricity: Lighting Garage (MBtu),0.0
+End Use: Electricity: Lighting Exterior (MBtu),0.0
+End Use: Electricity: Mech Vent (MBtu),0.95
+End Use: Electricity: Mech Vent Preheating (MBtu),0.0
+End Use: Electricity: Mech Vent Precooling (MBtu),0.0
+End Use: Electricity: Whole House Fan (MBtu),0.0
+End Use: Electricity: Refrigerator (MBtu),0.0
+End Use: Electricity: Freezer (MBtu),0.0
+End Use: Electricity: Dehumidifier (MBtu),10.55
+End Use: Electricity: Dishwasher (MBtu),0.0
+End Use: Electricity: Clothes Washer (MBtu),0.0
+End Use: Electricity: Clothes Dryer (MBtu),0.0
+End Use: Electricity: Range/Oven (MBtu),0.0
+End Use: Electricity: Ceiling Fan (MBtu),0.0
+End Use: Electricity: Television (MBtu),0.0
+End Use: Electricity: Plug Loads (MBtu),0.0
+End Use: Electricity: Electric Vehicle Charging (MBtu),0.0
+End Use: Electricity: Well Pump (MBtu),0.0
+End Use: Electricity: Pool Heater (MBtu),0.0
+End Use: Electricity: Pool Pump (MBtu),0.0
+End Use: Electricity: Hot Tub Heater (MBtu),0.0
+End Use: Electricity: Hot Tub Pump (MBtu),0.0
+End Use: Electricity: PV (MBtu),0.0
+End Use: Electricity: Generator (MBtu),0.0
+End Use: Natural Gas: Heating (MBtu),0.0
+End Use: Natural Gas: Hot Water (MBtu),0.0
+End Use: Natural Gas: Clothes Dryer (MBtu),0.0
+End Use: Natural Gas: Range/Oven (MBtu),0.0
+End Use: Natural Gas: Mech Vent Preheating (MBtu),0.0
+End Use: Natural Gas: Pool Heater (MBtu),0.0
+End Use: Natural Gas: Hot Tub Heater (MBtu),0.0
+End Use: Natural Gas: Grill (MBtu),0.0
+End Use: Natural Gas: Lighting (MBtu),0.0
+End Use: Natural Gas: Fireplace (MBtu),0.0
+End Use: Natural Gas: Generator (MBtu),0.0
+End Use: Fuel Oil: Heating (MBtu),1.21
+End Use: Fuel Oil: Hot Water (MBtu),0.0
+End Use: Fuel Oil: Clothes Dryer (MBtu),0.0
+End Use: Fuel Oil: Range/Oven (MBtu),0.0
+End Use: Fuel Oil: Mech Vent Preheating (MBtu),0.0
+End Use: Fuel Oil: Grill (MBtu),0.0
+End Use: Fuel Oil: Lighting (MBtu),0.0
+End Use: Fuel Oil: Fireplace (MBtu),0.0
+End Use: Propane: Heating (MBtu),0.0
+End Use: Propane: Hot Water (MBtu),0.0
+End Use: Propane: Clothes Dryer (MBtu),0.0
+End Use: Propane: Range/Oven (MBtu),0.0
+End Use: Propane: Mech Vent Preheating (MBtu),0.0
+End Use: Propane: Grill (MBtu),0.0
+End Use: Propane: Lighting (MBtu),0.0
+End Use: Propane: Fireplace (MBtu),0.0
+End Use: Propane: Generator (MBtu),0.0
+End Use: Wood Cord: Heating (MBtu),0.0
+End Use: Wood Cord: Hot Water (MBtu),0.0
+End Use: Wood Cord: Clothes Dryer (MBtu),0.0
+End Use: Wood Cord: Range/Oven (MBtu),0.0
+End Use: Wood Cord: Mech Vent Preheating (MBtu),0.0
+End Use: Wood Cord: Grill (MBtu),0.0
+End Use: Wood Cord: Lighting (MBtu),0.0
+End Use: Wood Cord: Fireplace (MBtu),0.0
+End Use: Wood Pellets: Heating (MBtu),0.0
+End Use: Wood Pellets: Hot Water (MBtu),0.0
+End Use: Wood Pellets: Clothes Dryer (MBtu),0.0
+End Use: Wood Pellets: Range/Oven (MBtu),0.0
+End Use: Wood Pellets: Mech Vent Preheating (MBtu),0.0
+End Use: Wood Pellets: Grill (MBtu),0.0
+End Use: Wood Pellets: Lighting (MBtu),0.0
+End Use: Wood Pellets: Fireplace (MBtu),0.0
+End Use: Coal: Heating (MBtu),0.0
+End Use: Coal: Hot Water (MBtu),0.0
+End Use: Coal: Clothes Dryer (MBtu),0.0
+End Use: Coal: Range/Oven (MBtu),0.0
+End Use: Coal: Mech Vent Preheating (MBtu),0.0
+End Use: Coal: Grill (MBtu),0.0
+End Use: Coal: Lighting (MBtu),0.0
+End Use: Coal: Fireplace (MBtu),0.0
+
+Load: Heating (MBtu),-1.22
+Load: Cooling (MBtu),28.18
+Load: Hot Water: Delivered (MBtu),0.0
+Load: Hot Water: Tank Losses (MBtu),2.53
+Load: Hot Water: Desuperheater (MBtu),0.0
+Load: Hot Water: Solar Thermal (MBtu),0.0
+
+Unmet Load: Heating (MBtu),0.01
+Unmet Load: Cooling (MBtu),6.7
+
+Peak Electricity: Winter Total (W),1483
+Peak Electricity: Summer Total (W),1792
+
+Peak Load: Heating (kBtu),7.98
+Peak Load: Cooling (kBtu),11.6
+
+Component Load: Heating: Roofs (MBtu),0.67
+Component Load: Heating: Ceilings (MBtu),0.0
+Component Load: Heating: Walls (MBtu),0.33
+Component Load: Heating: Rim Joists (MBtu),0.0
+Component Load: Heating: Foundation Walls (MBtu),0.0
+Component Load: Heating: Doors (MBtu),0.04
+Component Load: Heating: Windows (MBtu),0.0
+Component Load: Heating: Skylights (MBtu),0.0
+Component Load: Heating: Floors (MBtu),-0.02
+Component Load: Heating: Slabs (MBtu),0.0
+Component Load: Heating: Internal Mass (MBtu),-0.31
+Component Load: Heating: Infiltration (MBtu),0.13
+Component Load: Heating: Natural Ventilation (MBtu),0.0
+Component Load: Heating: Mechanical Ventilation (MBtu),1.74
+Component Load: Heating: Whole House Fan (MBtu),0.0
+Component Load: Heating: Ducts (MBtu),0.0
+Component Load: Heating: Internal Gains (MBtu),-3.8
+Component Load: Cooling: Roofs (MBtu),5.28
+Component Load: Cooling: Ceilings (MBtu),0.0
+Component Load: Cooling: Walls (MBtu),3.1
+Component Load: Cooling: Rim Joists (MBtu),0.0
+Component Load: Cooling: Foundation Walls (MBtu),0.0
+Component Load: Cooling: Doors (MBtu),0.66
+Component Load: Cooling: Windows (MBtu),0.0
+Component Load: Cooling: Skylights (MBtu),0.0
+Component Load: Cooling: Floors (MBtu),-0.02
+Component Load: Cooling: Slabs (MBtu),0.0
+Component Load: Cooling: Internal Mass (MBtu),-0.31
+Component Load: Cooling: Infiltration (MBtu),0.66
+Component Load: Cooling: Natural Ventilation (MBtu),0.0
+Component Load: Cooling: Mechanical Ventilation (MBtu),4.2
+Component Load: Cooling: Whole House Fan (MBtu),0.0
+Component Load: Cooling: Ducts (MBtu),0.0
+Component Load: Cooling: Internal Gains (MBtu),14.6
+
+Hot Water: Clothes Washer (gal),0
+Hot Water: Dishwasher (gal),0
+Hot Water: Fixtures (gal),0
+Hot Water: Distribution Waste (gal),0
