+++ test/test_samples_osw/run/run.log
@@ -1,0 +1,1 @@
+[16:28:09.226350 WARN] No valid weather file defined in either the osm or osw.
