+++ test/test_samples_osw/run/sqlite.err
@@ -1,0 +1,1 @@
+SQLite3 message, sqlite.err open for processing!
