+++ test/test_samples_osw/run/stdout-energyplus
@@ -1,0 +1,64 @@
+EnergyPlus Starting
+EnergyPlus, Version 9.4.0-998c4b761e, YMD=2021.02.08 16:28
+Adjusting Air System Sizing
+Adjusting Standard 62.1 Ventilation Sizing
+Initializing Simulation
+Reporting Surfaces
+Beginning Primary Simulation
+Initializing New Environment Parameters
+Warming up {1}
+Warming up {2}
+Warming up {3}
+Warming up {4}
+Warming up {5}
+Warming up {6}
+Warming up {7}
+Warming up {8}
+Warming up {9}
+Warming up {10}
+Warming up {11}
+Warming up {12}
+Warming up {13}
+Warming up {14}
+Warming up {15}
+Starting Simulation at 01/01/2007 for RUN PERIOD 1
+Updating Shadowing Calculations, Start Date=01/21/2007
+Continuing Simulation at 01/21/2007 for RUN PERIOD 1
+Updating Shadowing Calculations, Start Date=02/10/2007
+Continuing Simulation at 02/10/2007 for RUN PERIOD 1
+Updating Shadowing Calculations, Start Date=03/02/2007
+Continuing Simulation at 03/02/2007 for RUN PERIOD 1
+Updating Shadowing Calculations, Start Date=03/22/2007
+Continuing Simulation at 03/22/2007 for RUN PERIOD 1
+Updating Shadowing Calculations, Start Date=04/11/2007
+Continuing Simulation at 04/11/2007 for RUN PERIOD 1
+Updating Shadowing Calculations, Start Date=05/01/2007
+Continuing Simulation at 05/01/2007 for RUN PERIOD 1
+Updating Shadowing Calculations, Start Date=05/21/2007
+Continuing Simulation at 05/21/2007 for RUN PERIOD 1
+Updating Shadowing Calculations, Start Date=06/10/2007
+Continuing Simulation at 06/10/2007 for RUN PERIOD 1
+Updating Shadowing Calculations, Start Date=06/30/2007
+Continuing Simulation at 06/30/2007 for RUN PERIOD 1
+Updating Shadowing Calculations, Start Date=07/20/2007
+Continuing Simulation at 07/20/2007 for RUN PERIOD 1
+Updating Shadowing Calculations, Start Date=08/09/2007
+Continuing Simulation at 08/09/2007 for RUN PERIOD 1
+Updating Shadowing Calculations, Start Date=08/29/2007
+Continuing Simulation at 08/29/2007 for RUN PERIOD 1
+Updating Shadowing Calculations, Start Date=09/18/2007
+Continuing Simulation at 09/18/2007 for RUN PERIOD 1
+Updating Shadowing Calculations, Start Date=10/08/2007
+Continuing Simulation at 10/08/2007 for RUN PERIOD 1
+Updating Shadowing Calculations, Start Date=10/28/2007
+Continuing Simulation at 10/28/2007 for RUN PERIOD 1
+Updating Shadowing Calculations, Start Date=11/17/2007
+Continuing Simulation at 11/17/2007 for RUN PERIOD 1
+Updating Shadowing Calculations, Start Date=12/07/2007
+Continuing Simulation at 12/07/2007 for RUN PERIOD 1
+Updating Shadowing Calculations, Start Date=12/27/2007
+Continuing Simulation at 12/27/2007 for RUN PERIOD 1
+Computing Life Cycle Costs and Reporting
+Writing final SQL reports
+EnergyPlus Run Time=00hr 00min  6.80sec
+EnergyPlus Completed Successfully.
